PP-4352 Moved capture tests to relevant payment provider CaptureHandlerTest classes
