+++ src/test/java/uk/gov/pay/connector/gateway/smartpay/SmartpayPaymentProviderTest.java
@@ -1,12 +1,7 @@
 package uk.gov.pay.connector.gateway.smartpay;
 
-import com.codahale.metrics.Histogram;
-import com.codahale.metrics.MetricRegistry;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import fj.data.Either;
-import io.dropwizard.setup.Environment;
 import org.apache.commons.lang3.tuple.Pair;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
@@ -15,31 +10,19 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.pay.connector.app.ConnectorConfiguration;
-import uk.gov.pay.connector.app.GatewayConfig;
 import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
 import uk.gov.pay.connector.common.model.domain.Address;
-import uk.gov.pay.connector.gateway.ClientFactory;
-import uk.gov.pay.connector.gateway.GatewayClientFactory;
-import uk.gov.pay.connector.gateway.PaymentGatewayName;
 import uk.gov.pay.connector.gateway.model.Auth3dsDetails;
 import uk.gov.pay.connector.gateway.model.AuthCardDetails;
 import uk.gov.pay.connector.gateway.model.request.Auth3dsResponseGatewayRequest;
 import uk.gov.pay.connector.gateway.model.request.AuthorisationGatewayRequest;
-import uk.gov.pay.connector.gateway.model.request.CaptureGatewayRequest;
 import uk.gov.pay.connector.gateway.model.response.GatewayResponse;
-import uk.gov.pay.connector.gateway.worldpay.WorldpayCaptureResponse;
 import uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity;
 import uk.gov.pay.connector.usernotification.model.Notification;
 import uk.gov.pay.connector.usernotification.model.Notifications;
 import uk.gov.pay.connector.util.AuthUtils;
 import uk.gov.pay.connector.util.TestTemplateResourceLoader;
 
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.Entity;
-import javax.ws.rs.client.Invocation.Builder;
-import javax.ws.rs.client.WebTarget;
-import javax.ws.rs.core.Response;
 import java.io.IOException;
 
 import static io.dropwizard.testing.FixtureHelpers.fixture;
@@ -50,56 +33,24 @@ import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsSame.sameInstance;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 import static uk.gov.pay.connector.gateway.model.response.BaseAuthoriseResponse.AuthoriseStatus.AUTHORISED;
 import static uk.gov.pay.connector.gateway.model.response.BaseAuthoriseResponse.AuthoriseStatus.REQUIRES_3DS;
-import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity.Type.TEST;
 import static uk.gov.pay.connector.model.domain.ChargeEntityFixture.aValidChargeEntity;
 import static uk.gov.pay.connector.util.TestTemplateResourceLoader.SMARTPAY_AUTHORISATION_3DS_REQUIRED_RESPONSE;
 import static uk.gov.pay.connector.util.TestTemplateResourceLoader.SMARTPAY_AUTHORISATION_SUCCESS_RESPONSE;
 import static uk.gov.pay.connector.util.TestTemplateResourceLoader.SMARTPAY_CAPTURE_SUCCESS_RESPONSE;
 
 @RunWith(MockitoJUnitRunner.class)
-public class SmartpayPaymentProviderTest {
+public class SmartpayPaymentProviderTest extends BaseSmartpayPaymentProviderTest {
 
-    private SmartpayPaymentProvider provider;
-    private SmartpayCaptureHandler smartpayCaptureHandler;
-    private GatewayClientFactory gatewayClientFactory;
-
-    @Mock
-    private Client mockClient;
-    @Mock
-    private ClientFactory mockClientFactory;
     @Mock
     private GatewayAccountEntity mockGatewayAccountEntity;
-    @Mock
-    private ConnectorConfiguration configuration;
-    @Mock
-    private GatewayConfig gatewayConfig;
-    @Mock
-    private Environment environment;
-    @Mock
-    private MetricRegistry metricRegistry;
 
     @Before
     public void setup() {
-        gatewayClientFactory = new GatewayClientFactory(mockClientFactory);
-
-        when(mockClientFactory.createWithDropwizardClient(eq(PaymentGatewayName.SMARTPAY), any(MetricRegistry.class)))
-                .thenReturn(mockClient);
-        when(configuration.getGatewayConfigFor(PaymentGatewayName.SMARTPAY)).thenReturn(gatewayConfig);
-        when(gatewayConfig.getUrls()).thenReturn(ImmutableMap.of(TEST.toString(), "http://smartpay.url"));
-        when(environment.metrics()).thenReturn(metricRegistry);
-        when(metricRegistry.histogram(anyString())).thenReturn(mock(Histogram.class));
-
+        super.setup();
         mockSmartpaySuccessfulOrderSubmitResponse();
-
-        provider = new SmartpayPaymentProvider(configuration, gatewayClientFactory, environment, new ObjectMapper());
-        smartpayCaptureHandler = (SmartpayCaptureHandler) provider.getCaptureHandler();
     }
 
     @Test
@@ -181,19 +132,6 @@ public class SmartpayPaymentProviderTest {
     }
 
     @Test
-    public void shouldCaptureAPaymentSuccessfully() {
-
-        mockSmartpaySuccessfulCaptureResponse();
-
-        ChargeEntity chargeEntity = aValidChargeEntity()
-                .withGatewayAccountEntity(aServiceAccount())
-                .build();
-
-        GatewayResponse<SmartpayCaptureResponse> response = smartpayCaptureHandler.capture(CaptureGatewayRequest.valueOf(chargeEntity));
-        assertTrue(response.isSuccessful());
-    }
-
-    @Test
     public void parseNotification_shouldReturnErrorIfUnparseableSoapMessage() {
         Either<String, Notifications<Pair<String, Boolean>>> response = provider.parseNotification("not valid soap message");
         assertThat(response.isLeft(), is(true));
@@ -229,20 +167,6 @@ public class SmartpayPaymentProviderTest {
         assertThat(provider.verifyNotification(mock(Notification.class), mockGatewayAccountEntity), is(true));
     }
 
-    private GatewayAccountEntity aServiceAccount() {
-        GatewayAccountEntity gatewayAccount = new GatewayAccountEntity();
-        gatewayAccount.setId(1L);
-        gatewayAccount.setGatewayName("smartpay");
-        gatewayAccount.setCredentials(ImmutableMap.of(
-                "username", "theUsername",
-                "password", "thePassword",
-                "merchant_id", "theMerchantCode"
-        ));
-        gatewayAccount.setType(TEST);
-
-        return gatewayAccount;
-    }
-
     private void mockSmartpaySuccessfulOrderSubmitResponse() {
         mockSmartpayResponse(200, successAuthoriseResponse());
     }
@@ -257,20 +181,6 @@ public class SmartpayPaymentProviderTest {
 
     private void mockSmartpaySuccessfulCaptureResponse() {
         mockSmartpayResponse(200, successCaptureResponse());
-    }
-
-    private void mockSmartpayResponse(int httpStatus, String responsePayload) {
-        WebTarget mockTarget = mock(WebTarget.class);
-        when(mockClient.target(anyString())).thenReturn(mockTarget);
-        Builder mockBuilder = mock(Builder.class);
-        when(mockTarget.request()).thenReturn(mockBuilder);
-        when(mockBuilder.header(anyString(), any(Object.class))).thenReturn(mockBuilder);
-
-        Response response = mock(Response.class);
-        when(response.readEntity(String.class)).thenReturn(responsePayload);
-        when(mockBuilder.post(any(Entity.class))).thenReturn(response);
-
-        when(response.getStatus()).thenReturn(httpStatus);
     }
 
     private String successAuthoriseResponse() {
