+++ src/test/java/uk/gov/pay/connector/gateway/worldpay/WorldpayPaymentProviderTest.java
@@ -1,26 +1,16 @@
 package uk.gov.pay.connector.gateway.worldpay;
 
-import com.codahale.metrics.Counter;
-import com.codahale.metrics.Histogram;
-import com.codahale.metrics.MetricRegistry;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import fj.data.Either;
-import io.dropwizard.setup.Environment;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
-import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.pay.connector.app.ConnectorConfiguration;
-import uk.gov.pay.connector.app.GatewayConfig;
-import uk.gov.pay.connector.app.WorldpayConfig;
 import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
 import uk.gov.pay.connector.common.model.domain.Address;
-import uk.gov.pay.connector.gateway.CaptureHandler;
-import uk.gov.pay.connector.gateway.ClientFactory;
 import uk.gov.pay.connector.gateway.GatewayClient;
 import uk.gov.pay.connector.gateway.GatewayClientFactory;
 import uk.gov.pay.connector.gateway.GatewayOperation;
@@ -33,7 +23,6 @@ import uk.gov.pay.connector.gateway.model.GatewayError;
 import uk.gov.pay.connector.gateway.model.OrderRequestType;
 import uk.gov.pay.connector.gateway.model.request.Auth3dsResponseGatewayRequest;
 import uk.gov.pay.connector.gateway.model.request.AuthorisationGatewayRequest;
-import uk.gov.pay.connector.gateway.model.request.CaptureGatewayRequest;
 import uk.gov.pay.connector.gateway.model.request.RefundGatewayRequest;
 import uk.gov.pay.connector.gateway.model.response.GatewayResponse;
 import uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity;
@@ -45,15 +34,8 @@ import uk.gov.pay.connector.usernotification.model.Notifications;
 import uk.gov.pay.connector.util.AuthUtils;
 import uk.gov.pay.connector.util.TestTemplateResourceLoader;
 
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.Entity;
-import javax.ws.rs.client.Invocation;
-import javax.ws.rs.client.WebTarget;
-import javax.ws.rs.core.NewCookie;
-import javax.ws.rs.core.Response;
 import java.io.IOException;
 import java.time.ZonedDateTime;
-import java.util.Collections;
 import java.util.EnumMap;
 import java.util.Map;
 import java.util.function.BiFunction;
@@ -67,24 +49,16 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsSame.sameInstance;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.argThat;
 import static org.mockito.Mockito.eq;
 import static org.mockito.Mockito.isNull;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.pay.connector.gateway.model.ErrorType.GENERIC_GATEWAY_ERROR;
 import static uk.gov.pay.connector.gateway.model.ErrorType.UNEXPECTED_HTTP_STATUS_CODE_FROM_GATEWAY;
 import static uk.gov.pay.connector.gateway.model.GatewayError.unexpectedStatusCodeFromGateway;
-import static uk.gov.pay.connector.gateway.worldpay.WorldpayPaymentProvider.WORLDPAY_MACHINE_COOKIE_NAME;
-import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccount.CREDENTIALS_MERCHANT_ID;
-import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccount.CREDENTIALS_PASSWORD;
-import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccount.CREDENTIALS_USERNAME;
-import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity.Type.TEST;
 import static uk.gov.pay.connector.model.domain.ChargeEntityFixture.aValidChargeEntity;
 import static uk.gov.pay.connector.util.TestTemplateResourceLoader.WORLDPAY_NOTIFICATION;
 import static uk.gov.pay.connector.util.TestTemplateResourceLoader.WORLDPAY_VALID_3DS_RESPONSE_AUTH_WORLDPAY_REQUEST;
@@ -92,53 +66,12 @@ import static uk.gov.pay.connector.util.TestTemplateResourceLoader.WORLDPAY_VALI
 import static uk.gov.pay.connector.util.TestTemplateResourceLoader.WORLDPAY_VALID_AUTHORISE_WORLDPAY_REQUEST_INCLUDING_3DS;
 
 @RunWith(MockitoJUnitRunner.class)
-public class WorldpayPaymentProviderTest {
-
-    private WorldpayPaymentProvider provider;
-    private WorldpayCaptureHandler worldpayCaptureHandler;
-    private Map<String, String> urlMap = ImmutableMap.of(TEST.toString(), "http://worldpay.url");
-
-    EnumMap<GatewayOperation, GatewayClient> gatewayClients;
-
-    @Mock
-    private Client mockClient;
-    @Mock
-    private GatewayClient mockGatewayClient;
-    @Mock
-    private GatewayAccountEntity mockGatewayAccountEntity;
-    @Mock
-    private MetricRegistry mockMetricRegistry;
-    @Mock
-    private Histogram mockHistogram;
-    @Mock
-    private Counter mockCounter;
-    @Mock
-    private ClientFactory mockClientFactory;
-    @Mock
-    private ConnectorConfiguration configuration;
-    @Mock
-    private GatewayConfig gatewayConfig;
-    @Mock
-    private Environment environment;
+public class WorldpayPaymentProviderTest extends WorldpayBasePaymentProviderTest {
 
     @Before
     public void setup() {
-        GatewayClientFactory gatewayClientFactory = new GatewayClientFactory(mockClientFactory);
+        super.setup();
         mockWorldpaySuccessfulOrderSubmitResponse();
-        when(mockMetricRegistry.histogram(anyString())).thenReturn(mockHistogram);
-        when(mockMetricRegistry.counter(anyString())).thenReturn(mockCounter);
-        when(mockClientFactory.createWithDropwizardClient(
-                eq(PaymentGatewayName.WORLDPAY), any(GatewayOperation.class), any(MetricRegistry.class))
-        )
-                .thenReturn(mockClient);
-        
-        when(configuration.getGatewayConfigFor(PaymentGatewayName.WORLDPAY)).thenReturn(gatewayConfig);
-        when(gatewayConfig.getUrls()).thenReturn(urlMap);
-        when(configuration.getWorldpayConfig()).thenReturn(mock(WorldpayConfig.class));
-        when(environment.metrics()).thenReturn(mockMetricRegistry);
-
-        provider = new WorldpayPaymentProvider(configuration, gatewayClientFactory, environment);
-        worldpayCaptureHandler = (WorldpayCaptureHandler) provider.getCaptureHandler();
     }
 
     @Test
@@ -176,24 +109,24 @@ public class WorldpayPaymentProviderTest {
 
         GatewayClientFactory gatewayClientFactory = mock(GatewayClientFactory.class);
         when(gatewayClientFactory.createGatewayClient(any(PaymentGatewayName.class), any(GatewayOperation.class), any(Map.class), any(BiFunction.class), any())).thenReturn(mockGatewayClient);
-        
+
         WorldpayPaymentProvider worldpayPaymentProvider = new WorldpayPaymentProvider(configuration, gatewayClientFactory, environment);
         worldpayPaymentProvider.refund(RefundGatewayRequest.valueOf(refundEntity));
 
         String expectedRefundRequest =
                 "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" +
-                "<!DOCTYPE paymentService PUBLIC \"-//WorldPay//DTD WorldPay PaymentService v1//EN\"\n" +
-                "        \"http://dtd.worldpay.com/paymentService_v1.dtd\">\n" +
-                "<paymentService version=\"1.4\" merchantCode=\"MERCHANTCODE\">\n" +
-                "    <modify>\n" +
-                "        <orderModification orderCode=\"transaction-id\">\n" +
-                "            <refund reference=\"" +refundEntity.getExternalId()+ "\">\n" +
-                "                <amount currencyCode=\"GBP\" exponent=\"2\" value=\"500\"/>\n" +
-                "            </refund>\n" +
-                "        </orderModification>\n" +
-                "    </modify>\n" +
-                "</paymentService>\n" +
-                "";
+                        "<!DOCTYPE paymentService PUBLIC \"-//WorldPay//DTD WorldPay PaymentService v1//EN\"\n" +
+                        "        \"http://dtd.worldpay.com/paymentService_v1.dtd\">\n" +
+                        "<paymentService version=\"1.4\" merchantCode=\"MERCHANTCODE\">\n" +
+                        "    <modify>\n" +
+                        "        <orderModification orderCode=\"transaction-id\">\n" +
+                        "            <refund reference=\"" + refundEntity.getExternalId() + "\">\n" +
+                        "                <amount currencyCode=\"GBP\" exponent=\"2\" value=\"500\"/>\n" +
+                        "            </refund>\n" +
+                        "        </orderModification>\n" +
+                        "    </modify>\n" +
+                        "</paymentService>\n" +
+                        "";
 
         verify(mockGatewayClient).postRequestFor(eq(null), eq(mockGatewayAccountEntity), argThat(argument -> argument.getPayload().equals(expectedRefundRequest) && argument.getOrderRequestType().equals(OrderRequestType.REFUND)));
     }
@@ -281,7 +214,7 @@ public class WorldpayPaymentProviderTest {
 
         GatewayClientFactory gatewayClientFactory = mock(GatewayClientFactory.class);
         when(gatewayClientFactory.createGatewayClient(any(PaymentGatewayName.class), any(GatewayOperation.class), any(Map.class), any(BiFunction.class), any())).thenReturn(mockGatewayClient);
-        
+
         WorldpayPaymentProvider worldpayPaymentProvider = new WorldpayPaymentProvider(configuration, gatewayClientFactory, environment);
 
         worldpayPaymentProvider.authorise3dsResponse(get3dsResponseGatewayRequest(mockChargeEntity));
@@ -318,7 +251,7 @@ public class WorldpayPaymentProviderTest {
 
         GatewayClientFactory gatewayClientFactory = mock(GatewayClientFactory.class);
         when(gatewayClientFactory.createGatewayClient(any(PaymentGatewayName.class), any(GatewayOperation.class), any(Map.class), any(BiFunction.class), any())).thenReturn(mockGatewayClient);
-        
+
         WorldpayPaymentProvider worldpayPaymentProvider = new WorldpayPaymentProvider(configuration, gatewayClientFactory, environment);
 
         worldpayPaymentProvider.authorise3dsResponse(get3dsResponseGatewayRequest(mockChargeEntity));
@@ -339,14 +272,6 @@ public class WorldpayPaymentProviderTest {
     }
 
     @Test
-    public void shouldCaptureAPaymentSuccessfully() throws Exception {
-        mockWorldpaySuccessfulCaptureResponse();
-
-        GatewayResponse response = worldpayCaptureHandler.capture(getCaptureRequest());
-        assertTrue(response.isSuccessful());
-    }
-
-    @Test
     public void shouldErrorIfAuthorisationIsUnsuccessful() {
         mockWorldpayErrorResponse(401);
         GatewayResponse<WorldpayOrderStatusResponse> response = provider.authorise(getCardAuthorisationRequest());
@@ -355,26 +280,6 @@ public class WorldpayPaymentProviderTest {
         assertFalse(response.getSessionIdentifier().isPresent());
         assertThat(response.getGatewayError().isPresent(), is(true));
         assertEquals(response.getGatewayError().get(), new GatewayError("Unexpected HTTP status code 401 from gateway",
-                UNEXPECTED_HTTP_STATUS_CODE_FROM_GATEWAY));
-    }
-
-    @Test
-    public void shouldErrorIfOrderReferenceNotKnownInCapture() {
-        mockWorldpayErrorResponse(200);
-        GatewayResponse<WorldpayCaptureResponse> response = worldpayCaptureHandler.capture(getCaptureRequest());
-
-        assertThat(response.isFailed(), is(true));
-        assertThat(response.getGatewayError().isPresent(), is(true));
-        assertEquals(response.getGatewayError().get(), new GatewayError("Worldpay capture response (error code: 5, error: Order has already been paid)", GENERIC_GATEWAY_ERROR));
-    }
-
-    @Test
-    public void shouldErrorIfWorldpayResponseIsNot200() {
-        mockWorldpayErrorResponse(400);
-        GatewayResponse<WorldpayCaptureResponse> response = worldpayCaptureHandler.capture(getCaptureRequest());
-        assertThat(response.isFailed(), is(true));
-        assertThat(response.getGatewayError().isPresent(), is(true));
-        assertEquals(response.getGatewayError().get(), new GatewayError("Unexpected HTTP status code 400 from gateway",
                 UNEXPECTED_HTTP_STATUS_CODE_FROM_GATEWAY));
     }
 
@@ -451,90 +356,8 @@ public class WorldpayPaymentProviderTest {
         return getCardAuthorisationRequest(chargeEntity);
     }
 
-
-    private GatewayAccountEntity aServiceAccount() {
-        GatewayAccountEntity gatewayAccount = new GatewayAccountEntity();
-        gatewayAccount.setId(1L);
-        gatewayAccount.setGatewayName("worldpay");
-        gatewayAccount.setRequires3ds(false);
-        gatewayAccount.setCredentials(ImmutableMap.of(
-                CREDENTIALS_MERCHANT_ID, "worlpay-merchant",
-                CREDENTIALS_USERNAME, "worldpay-password",
-                CREDENTIALS_PASSWORD, "password"
-        ));
-        gatewayAccount.setType(TEST);
-        return gatewayAccount;
-    }
-
-    private void assertEquals(GatewayError actual, GatewayError expected) {
-        assertNotNull(actual);
-        assertThat(actual.getMessage(), is(expected.getMessage()));
-        assertThat(actual.getErrorType(), is(expected.getErrorType()));
-    }
-
-    private CaptureGatewayRequest getCaptureRequest() {
-        ChargeEntity chargeEntity = aValidChargeEntity()
-                .withGatewayAccountEntity(aServiceAccount())
-                .build();
-        return CaptureGatewayRequest.valueOf(chargeEntity);
-    }
-
-    private void mockWorldpayErrorResponse(int httpStatus) {
-        mockWorldpayResponse(httpStatus, errorResponse());
-    }
-
-    private void mockWorldpaySuccessfulCaptureResponse() {
-        mockWorldpayResponse(200, successCaptureResponse());
-    }
-
     private void mockWorldpaySuccessfulOrderSubmitResponse() {
         mockWorldpayResponse(200, successAuthoriseResponse());
-    }
-
-    private void mockWorldpayResponse(int httpStatus, String responsePayload) {
-        WebTarget mockTarget = mock(WebTarget.class);
-        when(mockClient.target(anyString())).thenReturn(mockTarget);
-        Invocation.Builder mockBuilder = mock(Invocation.Builder.class);
-        when(mockTarget.request()).thenReturn(mockBuilder);
-        when(mockBuilder.header(anyString(), any(Object.class))).thenReturn(mockBuilder);
-
-        Map<String, NewCookie> responseCookies =
-                Collections.singletonMap(WORLDPAY_MACHINE_COOKIE_NAME, NewCookie.valueOf("value-from-worldpay"));
-
-        Response response = mock(Response.class);
-        when(response.readEntity(String.class)).thenReturn(responsePayload);
-        when(mockBuilder.post(any(Entity.class))).thenReturn(response);
-        when(response.getCookies()).thenReturn(responseCookies);
-
-        when(response.getStatus()).thenReturn(httpStatus);
-    }
-
-    private String successCaptureResponse() {
-        return "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" +
-                "<!DOCTYPE paymentService PUBLIC \"-//WorldPay//DTD WorldPay PaymentService v1//EN\"\n" +
-                "        \"http://dtd.worldpay.com/paymentService_v1.dtd\">\n" +
-                "<paymentService version=\"1.4\" merchantCode=\"MERCHANTCODE\">\n" +
-                "    <reply>\n" +
-                "        <ok>\n" +
-                "            <captureReceived orderCode=\"MyUniqueTransactionId!\">\n" +
-                "                <amount value=\"500\" currencyCode=\"GBP\" exponent=\"2\" debitCreditIndicator=\"credit\"/>\n" +
-                "            </captureReceived>\n" +
-                "        </ok>\n" +
-                "    </reply>\n" +
-                "</paymentService>";
-    }
-
-    private String errorResponse() {
-        return "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" +
-                "<!DOCTYPE paymentService PUBLIC \"-//WorldPay//DTD WorldPay PaymentService v1//EN\"\n" +
-                "        \"http://dtd.worldpay.com/paymentService_v1.dtd\">\n" +
-                "<paymentService version=\"1.4\" merchantCode=\"MERCHANTCODE\">\n" +
-                "    <reply>\n" +
-                "        <error code=\"5\">\n" +
-                "            <![CDATA[Order has already been paid]]>\n" +
-                "        </error>\n" +
-                "    </reply>\n" +
-                "</paymentService>";
     }
 
     private String successAuthoriseResponse() {
